refactor!: remove Component file and related tests
BREAKING CHANGE: No longer exports Component class
